Fix tests and refactor
